+++ src/weitersager/argparser.py
@@ -11,12 +11,9 @@ Command line argument parsing
 from argparse import ArgumentParser
 from pathlib import Path
 
-from irc.bot import ServerSpec
-
 
 DEFAULT_HTTP_IP_ADDRESS = '127.0.0.1'
 DEFAULT_HTTP_PORT = 8080
-DEFAULT_IRC_PORT = 6667
 
 
 def create_parser():
@@ -24,32 +21,6 @@ def create_parser():
     parser = ArgumentParser()
 
     parser.add_argument('config_filename', type=Path, metavar='CONFIG_FILENAME')
-
-    parser.add_argument(
-        '--irc-nickname',
-        dest='irc_nickname',
-        default='Weitersager',
-        help='the IRC nickname the bot should use',
-        metavar='NICKNAME',
-    )
-
-    parser.add_argument(
-        '--irc-realname',
-        dest='irc_realname',
-        default='Weitersager',
-        help='the IRC realname the bot should use',
-        metavar='REALNAME',
-    )
-
-    parser.add_argument(
-        '--irc-server',
-        dest='irc_server',
-        type=parse_irc_server_arg,
-        help='IRC server (host and, optionally, port) to connect to '
-        + '[e.g. "irc.example.com" or "irc.example.com:6669"; '
-        + f'default port: {DEFAULT_IRC_PORT:d}]',
-        metavar='SERVER',
-    )
 
     parser.add_argument(
         '--http-listen-ip-address',
@@ -71,14 +42,6 @@ def create_parser():
     )
 
     return parser
-
-
-def parse_irc_server_arg(value):
-    """Parse a hostname with optional port."""
-    fragments = value.split(':', 1)
-    if len(fragments) > 1:
-        fragments[1] = int(fragments[1])
-    return ServerSpec(*fragments)
 
 
 def parse_args():