+++ src/weitersager/config.py
@@ -10,17 +10,44 @@ Configuration loading
 
 import rtoml
 
-from .irc import Channel
+from irc.bot import ServerSpec
+
+from .irc import Channel, Config as IrcConfig
 
 
-def load_irc_channels(path):
-    """Load IRC channel configuration from file."""
+def load_irc_config(path):
+    """Load IRC configuration from file."""
     data = rtoml.load(path)
-    return list(_get_irc_channels(data))
+
+    return _get_irc_config(data)
 
 
-def _get_irc_channels(data):
-    for channel in data['irc']['channels']:
+def _get_irc_config(data):
+    data_irc = data['irc']
+
+    server = _get_irc_server(data_irc)
+    nickname = data_irc['bot']['nickname']
+    realname = data_irc['bot']['realname']
+    channels = list(_get_irc_channels(data_irc))
+
+    return IrcConfig(
+        server=server,
+        nickname=nickname,
+        realname=realname,
+        channels=channels,
+    )
+
+
+def _get_irc_server(data_irc):
+    data_server = data_irc['server']
+    host = data_server['host']
+    port = int(data_server['port'])
+
+    return ServerSpec(host, port)
+
+
+def _get_irc_channels(data_irc):
+    for channel in data_irc['channels']:
         name = channel['name']
         password = channel.get('password')
         yield Channel(name, password)